+++ CHANGES.md
@@ -13,6 +13,8 @@
   Bifunctor instances for tuples of length higher than 2.
 * [#33](https://github.com/serokell/universum/issues/33):
   Add `($)` and `Each` type operators.
+* [#35](https://github.com/serokell/universum/issues/35):
+  Generalize `andM`, `orM`, `allM`, `anyM` over container type.
 
 0.2.1
 =====